1.[feat]: ads_date_topic table add payoff、payoff_user_count and payment_user_count columns 2.[refactor]: optimize dashboard header calculate

+++ litemall-report/src/main/java/org/tlh/dw/vo/DashBoardHeader.java
@@ -18,7 +18,7 @@ public class DashBoardHeader {
     private BigDecimal orderAmount;
     private BigDecimal paymentAmount;
     private BigDecimal refundAmount;
-    private BigDecimal payConvertRate;
-    private double prePrice;
+    private double payConvertRate;
+    private BigDecimal prePrice;
 
 }
